+++ src/app/events/VoiceStateUpdate.ts
@@ -1,7 +1,7 @@
 import { VoiceState } from 'discord.js';
 import addUserForEvent from './poap/AddUserForEvent';
 import { DiscordEvent } from '../types/discord/DiscordEvent';
-import ServiceUtils from '../utils/ServiceUtils';
+import { LogUtils } from '../utils/Log';
 
 /**
  * voiceStateUpdate
@@ -18,11 +18,9 @@ export default class implements DiscordEvent {
 	 */
 	async execute(oldState: VoiceState, newState: VoiceState): Promise<any> {
 		try {
-			if (ServiceUtils.isBanklessDAO(oldState.guild)) {
-				await addUserForEvent(oldState, newState).catch(console.error);
-			}
+			await addUserForEvent(oldState, newState).catch(e => LogUtils.logError('failed to add user for POAP event', e, oldState.guild.id));
 		} catch (e) {
-			console.error(e);
+			LogUtils.logError('failed to process event voiceStateUpdate', e);
 		}
 	}
 }