+++ src/app/events/bounty/MessageReactionAddBounty.ts
@@ -9,6 +9,7 @@ import BountyUtils from '../../utils/BountyUtils';
 import RefreshBounty from '../../service/bounty/RefreshBounty';
 import UpdateEditKeyBounty from '../../service/bounty/UpdateEditKeyBounty';
 import ReCreateBounty from '../../service/bounty/ReCreateBounty';
+import Log, { LogUtils } from '../../utils/Log';
 
 export default async (reaction: MessageReaction, user: User): Promise<any> | null => {
 	if (reaction.message.channel.id !== channelIds.bountyBoard) {
@@ -25,36 +26,36 @@ export default async (reaction: MessageReaction, user: User): Promise<any> | nul
 	const guildMember: GuildMember = await reaction.message.guild.members.fetch(user);
 	
 	if (message.webhookId !== null) {
-		console.log('message created by webhook');
+		Log.info('message created by webhook');
 		await message.delete();
-		message = await ReCreateBounty(guildMember, bountyId).catch(console.error) as Message;
+		message = await ReCreateBounty(guildMember, bountyId).catch(e => LogUtils.logError('recreating bounty failed', e)) as Message;
 	}
 	
 	if (message === null) {
-		console.log('message not found');
+		Log.debug('message not found');
 		return;
 	}
 
 	if (reaction.emoji.name === '🏴') {
-		console.log(`${user.tag} attempting to claim a bounty ${bountyId} from the bounty board`);
-		return claimBountyForValidId(guildMember, bountyId, message).catch(console.error);
+		Log.info(`${user.tag} attempting to claim a bounty ${bountyId} from the bounty board`);
+		return claimBountyForValidId(guildMember, bountyId, message).catch(e => LogUtils.logError('failed to claim bounty', e));
 	} else if (reaction.emoji.name === '📝') {
-		return UpdateEditKeyBounty(guildMember, bountyId, message).catch(console.error);
+		return UpdateEditKeyBounty(guildMember, bountyId, message).catch(e => LogUtils.logError('failed to update bounty', e));
 	} else if (reaction.emoji.name === '❌') {
-		console.log(`${user.tag} attempting to delete bounty ${bountyId}`);
-		return deleteBountyForValidId(guildMember, bountyId, message).catch(console.error);
+		Log.info(`${user.tag} attempting to delete bounty ${bountyId}`);
+		return deleteBountyForValidId(guildMember, bountyId, message).catch(e => LogUtils.logError('failed to delete bounty', e));
 	} else if (reaction.emoji.name === '📮') {
-		console.log(`${user.tag} attempting to submit bounty ${bountyId}`);
+		Log.info(`${user.tag} attempting to submit bounty ${bountyId}`);
 		// TODO: have bot ask user for details
-		return submitBountyForValidId(guildMember, bountyId, null, null, message).catch(console.error);
+		return submitBountyForValidId(guildMember, bountyId, null, null, message).catch(e => LogUtils.logError('failed to submit bounty', e));
 	} else if (reaction.emoji.name === '✅') {
-		console.log(`${user.tag} attempting to mark bounty ${bountyId} complete`);
-		return completeBountyForValidId(guildMember, bountyId, message).catch(console.error);
+		Log.info(`${user.tag} attempting to mark bounty ${bountyId} complete`);
+		return completeBountyForValidId(guildMember, bountyId, message).catch(e => LogUtils.logError('failed to complete bounty', e));
 	} else if (reaction.emoji.name === '🆘') {
-		console.log(`${user.tag} attempting to seek help for bounty ${bountyId}`);
-		return seekHelpValidBountyId(guildMember, bountyId).catch(console.error);
+		Log.info(`${user.tag} attempting to seek help for bounty ${bountyId}`);
+		return seekHelpValidBountyId(guildMember, bountyId).catch(e => LogUtils.logError('failed to seek help for bounty', e));
 	} else if (reaction.emoji.name === '🔄') {
-		console.log(`${user.tag} attempting to refresh bounty ${bountyId}`);
-		return RefreshBounty(guildMember, bountyId, message).catch(console.error);
+		Log.info(`${user.tag} attempting to refresh bounty ${bountyId}`);
+		return RefreshBounty(guildMember, bountyId, message).catch(e => LogUtils.logError('failed to refresh bounty', e));
 	}
 };